build: memory-safety wave — ingestion sanitizer, CI, libc++ hardening, bounds-by-construction DSP
Untrusted-input boundary normalization:
- New Source/dsp/patch_sanitizer.h normalizes raw Patch (112 B) and
  PartData (84 B) bytes to their firmware-valid domains AT INGESTION,
  wired at the two paths that previously pushed file/host bytes into
  the engine unvalidated: .MUL multi loads
  (ParvatiAudioProcessor::loadMultiFile) and host-state blob restores
  (SynthEngine::restoreState). Bounds derive from the patch.h enums
  (plus a static_assert against kNumTuningPresets) so they cannot drift
  from the DSP definitions. IDENTITY for legitimate firmware files
  (round-trips byte-exact — pinned by roundtrip/golden/multi_load/
  host_state/loader_fuzz); deterministically narrows hostile blobs at
  every sink at once. The 2026-08-18 DSP-side sink clamps stay
  (defense in depth). Regression cover: tests/patch_sanitizer_test.cpp
  (unit domains + identity + end-to-end wiring proofs through a real
  written .MUL and a real captured/restored state blob).

Bounds-by-construction DSP migration (zero runtime cost, bit-identical
audio — validated by firmware_parity / roundtrip_golden / the full
unified suite):
- FV1 fx processors: setParams(const float[5]) -> std::array<float,
  kNumFxSlotParams> across all 15 Fv1* processors + FxChain/FxProcessor.
- dsp::Voice: sized std::array output block, bounds-guarded mod-matrix
  byte accessors, clamped envelope indexing, removed the unguarded
  mutable_patch_data() flat alias; AmbikaVoice consumers updated to
  the sized-array contract (crush shadow included).
- tests/dsp_memory_safety_test.cpp pins the guard behaviors.

libc++ hardened mode FAST (_LIBCPP_HARDENING_MODE) active in
RelWithDebInfo/Release (skipped under sanitizers/Debug): std::array/
std::vector bounds checks the migration leans on stay live in shipped
artifacts. Near-zero overhead; inert macro on older toolchains.

GitHub Actions CI (.github/workflows/ci.yml): PR gate (tests-only
Debug build + full unified suite minus loader_fuzz/perf_smoke) and
nightly (full 119-test native suite + Release Standalone/VST3/CLAP
proving the format wrappers and the Release-only libc++ hardening
configure + ASan/UBSan sweep via tools/run_sanitizers.sh + TSan on
the concurrency-critical subset). External deps cloned per run
(ambika firmware reference — untracked per NOTICES.md — and JUCE
9.0.1).

Full unified suite with this tree: 119/119.

+++ Source/dsp/clouds/clouds/dsp/correlator.cc
@@ -57,7 +57,13 @@ void Correlator::EvaluateNextCandidate() {
     uint32_t source_bits = source[i];
     uint32_t destination_bits = 0;
     destination_bits |= destination[i] << offset_bits;
-    destination_bits |= destination[i + 1] >> (32 - offset_bits);
+    // Parvati memory-safety fix: upstream shifts a 32-bit word right by
+    // (32 - offset_bits), which is >> 32 (UB) when offset_bits == 0. On the
+    // firmware's ARM the lsr instruction maps that to 0; widen to 64-bit so
+    // the portable build reproduces exactly that result. In-bounds shifts
+    // are bit-identical to upstream.
+    destination_bits |= static_cast<uint32_t>(
+        static_cast<uint64_t>(destination[i + 1]) >> (32 - offset_bits));
     uint32_t count = ~(source_bits ^ destination_bits);
     count = count - ((count >> 1) & 0x55555555);
     count = (count & 0x33333333) + ((count >> 2) & 0x33333333);
